Add files via upload
Here is the basic bones of the site. All is up for discussion and/or changes. I just wanted everyone to have the structure to start with. I created 5 pages we committed to and the option (if we have time or inclination) of the 6th page we discussed for the blog. I have an image file for the images I manipulated and the logo I created and the other photos, as you will see, are just the image addresses. I left the cards on the index blank deferring to Mike who is taking the database task and left the carousel placement (index or gallery page) up for discussion among us.